pyproject.toml
[tool.poetry]
name = "ml-cookiecutter-tabular"
version = "0.1.0"
description = "End-to-end ML template for tabular classification (DVC + MLflow + Hydra + FastAPI)"
authors = ["Samuel Drotar <sam.drotar@gmail.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.2"
numpy = "^2.0.0"
scikit-learn = "^1.5.0"
lightgbm = "^4.5.0"
xgboost = "^2.1.0"
mlflow = "^2.13.0"
dvc = "^3.50.2"
dvc-s3 = "^3.0.1"
hydra-core = "^1.3.2"
omegaconf = "^2.3.0"
fastapi = "^0.112.0"
uvicorn = "^0.30.0"
pydantic = "^2.8.2"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.1"
shap = "^0.46.0"
matplotlib = "^3.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-cov = "^5.0.0"
black = "^24.8.0"
isort = "^5.13.2"
flake8 = "^7.1.0"
mypy = "^1.11.1"
pre-commit = "^3.8.0"
httpx = "^0.27.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-q --maxfail=1 --disable-warnings --cov=src"
